feat(regional): Split regional analysis into its own tab
