+++ lib/actions/analysis/regional.js
@@ -12,8 +12,9 @@ import {REGIONAL_ANALYSIS_SAMPLING_DISTRIBUTION_COMPONENT} from '../../constants
 import {serverAction} from '../network'
 import {addComponent} from '../map'
 import authenticatedFetch from '../../utils/authenticated-fetch'
+import convertToR5Modification from '../../utils/convert-to-r5-modification'
 
-import type {LonLat, RegionalAnalysis} from '../../types'
+import type {Bounds, LonLat} from '../../types'
 
 export const setRegionalAnalyses = createAction('set regional analyses')
 
@@ -139,14 +140,62 @@ export const loadRegionalAnalysisGrids = async ({
   }
 }
 
-export const runAnalysis = (analysis: RegionalAnalysis) =>
-  serverAction({
-    url: `/api/regional`,
+export const createRegionalAnalysis = ({
+  bounds,
+  name,
+  profileRequest,
+  travelTimePercentile,
+  variantId
+}: {
+  bounds: Bounds,
+  name: string,
+  profileRequest: any,
+  travelTimePercentile: number,
+  variantId: number
+}) => (dispatch: () => void, getState: () => any) => {
+  const {analysis, project, scenario} = getState()
+  const {north, south, east, west} = bounds
+  const {currentIndicator, isochroneCutoff} = analysis
+  const {currentScenario} = scenario
+
+  dispatch(fetch({
     options: {
       method: 'POST',
-      body: JSON.stringify(analysis)
-    }
-  })
+      body: JSON.stringify({
+        name,
+        workerVersion: project.currentProject.r5version,
+        bundleId: currentScenario.bundleId,
+        grid: currentIndicator,
+        cutoffMinutes: isochroneCutoff,
+        travelTimePercentile,
+        scenarioId: currentScenario.id,
+        variant: variantId,
+        bounds: {
+          type: 'Polygon',
+          coordinates: [
+            [
+              [west, north],
+              [east, north],
+              [east, south],
+              [west, south],
+              [west, north]
+            ]
+          ]
+        },
+        request: {
+          ...profileRequest,
+          scenario: {
+            modifications: scenario.modifications
+              .filter(m => m.variants[variantId])
+              .map(convertToR5Modification)
+          },
+          type: 'REGIONAL_ANALYSIS'
+        }
+      })
+    },
+    url: '/api/regional'
+  }))
+}
 
 const deleteRegionalAnalysisLocally = createAction('delete regional analysis')
 export const deleteRegionalAnalysis = (analysisId: string) => [
