+++ lib/components/__tests__/modal.js
@@ -1,10 +1,11 @@
 /* global describe, it, expect, jest */
 
-describe('Component > Modal', () => {
-  const renderer = require('react-test-renderer')
-  const React = require('react')
-  const Modal = require('../modal')
+import renderer from 'react-test-renderer'
+import React from 'react'
 
+import Modal from '../modal'
+
+describe('Component > Modal', () => {
   it('renders correctly', () => {
     const onRequestCloseFn = jest.fn()
     renderer.create(
