+++ lib/components/analysis/index.js
@@ -8,11 +8,9 @@ import {sprintf} from 'sprintf-js'
 
 import BookmarkChooser from '../../containers/bookmark-chooser'
 import messages from '../../utils/messages'
-import Regional from './regional-analysis-selector'
-import convertToR5Modification from '../../utils/convert-to-r5-modification'
 import {Button} from '../buttons'
 import StackedPercentileSelector from './stacked-percentile-selector'
-import {Group, Text, Slider} from '../input'
+import {Group, Slider} from '../input'
 import Collapsible from '../collapsible'
 import ProfileRequestEditor from './profile-request-editor'
 import ScenarioApplicationError from './scenario-application-error'
@@ -57,13 +55,13 @@ type Props = {
   profileRequest: ProfileRequest,
   projectId: string,
   projectBounds: Bounds,
-  regionalAnalyses?: Array<Bounds & {
-      id: string,
-      name: string,
-      height: number,
-      width: number,
-      zoom: number
-    }>,
+  regionalAnalyses: Array<Bounds & {
+    id: string,
+    name: string,
+    height: number,
+    width: number,
+    zoom: number
+  }>,
   regionalAnalysisBounds?: Bounds,
   scenarioApplicationErrors?: Array<{}>,
   scenarioApplicationWarnings?: Array<{}>,
@@ -85,17 +83,14 @@ type Props = {
   addOpportunityLayerToMap(): void,
   clearIsochroneResults(): void,
   clearComparisonScenario(): void,
-  deleteRegionalAnalysis(): void,
+  createRegionalAnalysis: (any) => void,
   enterAnalysisMode(): void,
   exitAnalysisMode(): void,
   fetchTravelTimeSurface(): void,
-  loadRegionalAnalyses(): void,
   push(string): void,
   removeEditRegionalAnalysisBoundsLayerFromMap(): void,
   removeIsochroneLayerFromMap(): void,
   removeOpportunityLayerFromMap(): void,
-  runAnalysis(): void,
-  selectRegionalAnalysis(): void,
   setActiveVariant(): void,
   setComparisonScenario(): void,
   setCurrentIndicator(): void,
@@ -106,6 +101,7 @@ type Props = {
 
 type State = {
   creatingRegionalAnalysis: boolean,
+  editingBounds: boolean,
   regionalAnalysisPercentile: number,
   regionalAnalysisName: void | string
 }
@@ -116,6 +112,7 @@ type State = {
 export default class SinglePointAnalysis extends Component<void, Props, State> {
   state = {
     creatingRegionalAnalysis: false,
+    editingBounds: false,
     regionalAnalysisPercentile: 50,
     regionalAnalysisName: undefined
   }
@@ -156,62 +153,25 @@ export default class SinglePointAnalysis extends Component<void, Props, State> {
     if (e) setCurrentIndicator(projectId, e.value)
   }
 
-  /** run a new query */
-  runAnalysis = () => {
-    const {
-      bundleId,
-      currentIndicator,
-      isochroneCutoff,
-      modifications,
-      profileRequest,
-      regionalAnalysisBounds,
-      removeEditRegionalAnalysisBoundsLayerFromMap,
-      runAnalysis,
-      scenarioId,
-      variantIndex,
-      workerVersion
-    } = this.props
+  _createRegionalAnalysis = () => {
+    const {regionalAnalyses} = this.props
+    const name = window.prompt('To begin a regional analysis job for this scenario and settings enter a name and click ok.', `Analysis ${regionalAnalyses.length + 1}`)
+    if (name && name.length > 0) {
+      const {
+        createRegionalAnalysis,
+        profileRequest,
+        projectBounds,
+        regionalAnalysisBounds,
+        variantIndex
+      } = this.props
+      const {regionalAnalysisPercentile} = this.state
 
-    const {regionalAnalysisName, regionalAnalysisPercentile} = this.state
-
-    removeEditRegionalAnalysisBoundsLayerFromMap()
-
-    if (regionalAnalysisBounds) {
-      const {north, south, east, west} = regionalAnalysisBounds
-
-      runAnalysis({
-        name: regionalAnalysisName,
-        workerVersion,
-        bundleId,
-        grid: currentIndicator,
-        cutoffMinutes: isochroneCutoff,
+      createRegionalAnalysis({
+        bounds: regionalAnalysisBounds || projectBounds,
+        name,
+        profileRequest,
         travelTimePercentile: regionalAnalysisPercentile,
-        scenarioId,
-        variant: variantIndex,
-        bounds: {
-          type: 'Polygon',
-          coordinates: [
-            [
-              [west, north],
-              [east, north],
-              [east, south],
-              [west, south],
-              [west, north]
-            ]
-          ]
-        },
-        request: {
-          ...profileRequest,
-          scenario: {
-            modifications: modifications.map(convertToR5Modification)
-          },
-          type: 'REGIONAL_ANALYSIS'
-        }
-      })
-
-      this.setState({
-        regionalAnalysisName: undefined,
-        creatingRegionalAnalysis: false
+        variantId: variantIndex
       })
     }
   }
@@ -308,6 +268,22 @@ export default class SinglePointAnalysis extends Component<void, Props, State> {
     setRegionalAnalysisBounds(projectBounds)
     addEditRegionalAnalysisBoundsLayerToMap()
     this.setState({creatingRegionalAnalysis: true})
+  }
+
+  _showBoundsSelector = () => {
+    const {
+      setRegionalAnalysisBounds,
+      projectBounds,
+      addEditRegionalAnalysisBoundsLayerToMap
+    } = this.props
+    setRegionalAnalysisBounds(projectBounds)
+    addEditRegionalAnalysisBoundsLayerToMap()
+    this.setState({editingBounds: true})
+  }
+
+  _hideBoundsSelector = () => {
+    this.props.removeEditRegionalAnalysisBoundsLayerFromMap()
+    this.setState({editingBounds: false})
   }
 
   setRegionalAnalysisName = (e: Event & {target: HTMLInputElement}) =>
@@ -417,18 +393,12 @@ export default class SinglePointAnalysis extends Component<void, Props, State> {
     const {
       accessibility,
       currentIndicator,
-      deleteRegionalAnalysis,
       indicators,
       isFetchingIsochrone,
       isochroneFetchStatusMessage,
-      loadRegionalAnalyses,
       profileRequest,
-      projectId,
-      regionalAnalyses,
       scenarioApplicationErrors,
       scenarioApplicationWarnings,
-      scenarioId,
-      selectRegionalAnalysis,
       setProfileRequest,
       percentileCurves,
       variantIndex,
@@ -436,7 +406,6 @@ export default class SinglePointAnalysis extends Component<void, Props, State> {
     } = this.props
 
     const {
-      regionalAnalysisName,
       regionalAnalysisPercentile,
       creatingRegionalAnalysis
     } = this.state
@@ -496,14 +465,6 @@ export default class SinglePointAnalysis extends Component<void, Props, State> {
             />
           </Group>
 
-          <Collapsible title={messages.analysis.settings}>
-            <BookmarkChooser />
-            <ProfileRequestEditor
-              profileRequest={profileRequest}
-              setProfileRequest={setProfileRequest}
-            />
-          </Collapsible>
-
           {isFetchingIsochrone &&
             <div className='alert alert-info'>
               <Icon type='refresh' className={'fa-spin'} />{' '}
@@ -521,6 +482,21 @@ export default class SinglePointAnalysis extends Component<void, Props, State> {
                 <Icon type='refresh' /> {messages.analysis.refresh}
               </Button>
             </Group>}
+
+          {!isFetchingIsochrone &&
+            !scenarioApplicationErrors &&
+            percentileCurves &&
+            !creatingRegionalAnalysis &&
+            <Group>
+              <Button
+                style='success'
+                block
+                onClick={this._createRegionalAnalysis}
+              >
+                <Icon type='plus' />
+                {messages.analysis.newRegionalAnalysis}
+              </Button>
+            </Group>}
         </div>
 
         {accessibility != null &&
@@ -528,72 +504,33 @@ export default class SinglePointAnalysis extends Component<void, Props, State> {
           percentileCurves &&
           this.renderResults()}
 
-        <div className='DockTitle'>
-          {messages.analysis.regionalAnalyses}
-        </div>
-        <div className='block'>
-          {percentileCurves &&
-            !creatingRegionalAnalysis &&
-            <Group>
-              <Button
-                style='success'
-                block
-                onClick={this.prepareCreateRegionalAnalysis}
-              >
-                <Icon type='plus' />
-                {messages.analysis.newRegionalAnalysis}
-              </Button>
-            </Group>}
+        {!isFetchingIsochrone &&
+          !scenarioApplicationErrors &&
+          <div className='block'>
+            <BookmarkChooser />
+            <ProfileRequestEditor
+              profileRequest={profileRequest}
+              setProfileRequest={setProfileRequest}
+            />
 
-          {percentileCurves &&
-            creatingRegionalAnalysis &&
-            <form>
-              <Text
-                value={regionalAnalysisName}
-                label={messages.analysis.regionalAnalysisName}
-                onChange={this.setRegionalAnalysisName}
-              />
-              {this.renderRegionalAnalysisBoundsSelector()}
+            {this.renderRegionalAnalysisBoundsSelector()}
 
-              {/*
-                * A slider for selecting the percentile to use in a regional analysis.
-                * This will eventually be moved into analysis settings once we can calculate multiple
-                * percentiles in single point mode.
-                */}
-              <Slider
-                value={regionalAnalysisPercentile}
-                output
-                format='.1f'
-                min={1}
-                max={99}
-                step={1}
-                label={messages.analysis.travelTimePercentile}
-                onChange={this.setRegionalAnalysisPercentile}
-              />
-
-              <Button
-                block
-                type='submit'
-                style='success'
-                onClick={this.runAnalysis}
-                disabled={
-                  regionalAnalysisName == null || regionalAnalysisName === ''
-                }
-                aria-label={messages.analysis.startRegionalAnalysis}
-              >
-                <Icon type='check' /> {messages.analysis.startRegionalAnalysis}
-              </Button>
-            </form>}
-
-          <Regional
-            scenarioId={scenarioId}
-            projectId={projectId}
-            regionalAnalyses={regionalAnalyses}
-            loadRegionalAnalyses={loadRegionalAnalyses}
-            selectRegionalAnalysis={selectRegionalAnalysis}
-            deleteRegionalAnalysis={deleteRegionalAnalysis}
-          />
-        </div>
+            {/*
+              * A slider for selecting the percentile to use in a regional analysis.
+              * This will eventually be moved into analysis settings once we can calculate multiple
+              * percentiles in single point mode.
+              */}
+            <Slider
+              value={regionalAnalysisPercentile}
+              output
+              format='.1f'
+              min={1}
+              max={99}
+              step={1}
+              label={`${messages.analysis.travelTimePercentile} -- Regional analysis only`}
+              onChange={this.setRegionalAnalysisPercentile}
+            />
+          </div>}
       </div>
     )
   }
@@ -602,11 +539,9 @@ export default class SinglePointAnalysis extends Component<void, Props, State> {
   renderRegionalAnalysisBoundsSelector () {
     const {regionalAnalysisBounds, projectBounds, regionalAnalyses} = this.props
 
-    if (!regionalAnalysisBounds || !regionalAnalyses) return null
-
     // figure out which is selected
     let selected
-    if (boundsEqual(regionalAnalysisBounds, projectBounds)) {
+    if (!regionalAnalysisBounds || boundsEqual(regionalAnalysisBounds, projectBounds)) {
       selected = '__PROJECT' // special value
     } else {
       const selectedAnalysis = regionalAnalyses.find(r =>
@@ -634,6 +569,8 @@ export default class SinglePointAnalysis extends Component<void, Props, State> {
       })
     }
 
+    const {editingBounds} = this.state
+
     return (
       <Group label={messages.analysis.regionalBounds}>
         <Select
@@ -641,6 +578,20 @@ export default class SinglePointAnalysis extends Component<void, Props, State> {
           options={options}
           onChange={this.setRegionalAnalysisBounds}
         />
+        <br />
+        {editingBounds
+          ? <Button
+            block
+            onClick={this._hideBoundsSelector}
+            style='warning'
+            ><Icon type='stop' /> Stop editing bounds
+          </Button>
+          : <Button
+            block
+            onClick={this._showBoundsSelector}
+            style='warning'
+          ><Icon type='pencil' /> Set custom geographic bounds
+        </Button>}
       </Group>
     )
   }
