+++ lib/components/analysis/regional-analysis-selector.js
@@ -1,75 +1,52 @@
-import Icon from '@conveyal/woonerf/components/icon'
-import React, {PureComponent} from 'react'
+// @flow
+import memoize from 'lodash/memoize'
 import sortBy from 'lodash/sortBy'
-import PropTypes from 'prop-types'
+import React, {Component} from 'react'
 
 import messages from '../../utils/messages'
 
-const REFETCH_INTERVAL = 5 * 1000
+type Props = {
+  allAnalyses: any[],
+
+  goToAnalysis: (analysisId: string) => void
+}
 
 /**
  * Show progress of regional analysis, and allow displaying a regional analysis
  * on the map
  */
-export default class RegionalAnalysisSelector extends PureComponent {
-  static propTypes = {
-    loadRegionalAnalyses: PropTypes.func.isRequired,
-    projectId: PropTypes.string.isRequired,
-    regionalAnalyses: PropTypes.array,
-    scenarioId: PropTypes.string.isRequired,
-    selectRegionalAnalysis: PropTypes.func.isRequired,
-    deleteRegionalAnalysis: PropTypes.func.isRequired
-  }
+export default class RegionalAnalysisSelector extends Component {
+  props: Props
+  interval: ?number
 
   state = {
-    sortedRegionalAnalyses: sortByName(this.props.regionalAnalyses || [])
+    sortedAnalyses: sortByName(this.props.allAnalyses || [])
   }
 
-  componentWillMount () {
-    this.fetch()
-    this.interval = window.setInterval(this.fetch, REFETCH_INTERVAL)
-  }
-
-  componentWillUnmount () {
-    window.clearInterval(this.interval)
-  }
-
-  componentWillReceiveProps (nextProps) {
-    const {regionalAnalyses} = nextProps
-    if (regionalAnalyses !== this.props.regionalAnalyses) {
+  componentWillReceiveProps (nextProps: Props) {
+    const {allAnalyses} = nextProps
+    if (allAnalyses !== this.props.allAnalyses) {
       this.setState({
-        sortedRegionalAnalyses: sortByName(regionalAnalyses || [])
+        sortedAnalyses: sortByName(allAnalyses || [])
       })
     }
   }
 
-  fetch = () => {
-    const {loadRegionalAnalyses, projectId} = this.props
-    loadRegionalAnalyses(projectId)
-  }
-
-  _selectRegionalAnalysis = analysisId => {
-    const {projectId, scenarioId, selectRegionalAnalysis} = this.props
-    selectRegionalAnalysis(projectId, scenarioId, analysisId)
-  }
-
-  _deleteRegionalAnalysis = analysisId => {
-    const {deleteRegionalAnalysis} = this.props
-    deleteRegionalAnalysis(analysisId)
-  }
+  _selectRegionalAnalysis = memoize((analysisId: string) => () => {
+    this.props.goToAnalysis(analysisId)
+  })
 
   render () {
-    const {sortedRegionalAnalyses} = this.state
+    const {sortedAnalyses} = this.state
     return (
       <ul className='list-group'>
-        {sortedRegionalAnalyses
-          .filter(a => !a.deleted)
+        {sortedAnalyses
+          .filter(a => !a.deleted) // TODO: don't load the deleted ones...
           .map(a =>
             <Analysis
               analysis={a}
               key={`analysis-${a.id}`}
-              onSelect={this._selectRegionalAnalysis}
-              deleteRegionalAnalysis={this._deleteRegionalAnalysis}
+              onSelect={this._selectRegionalAnalysis(a.id)}
             />
           )}
       </ul>
@@ -77,7 +54,7 @@ export default class RegionalAnalysisSelector extends PureComponent {
   }
 }
 
-function Analysis ({analysis, onSelect, deleteRegionalAnalysis}) {
+function Analysis ({analysis, onSelect}) {
   const complete =
     analysis.status == null ||
     analysis.status.complete === analysis.status.total
@@ -86,8 +63,15 @@ function Analysis ({analysis, onSelect, deleteRegionalAnalysis}) {
     : analysis.status.complete / analysis.status.total * 100
 
   return (
-    <li className='list-group-item' key={`analysis-selector-${analysis.id}`}>
-      {analysis.name}
+    <li className='list-group-item'>
+      {complete
+        ? <a
+          onClick={onSelect}
+          tabIndex={0}
+          title={messages.analysis.openRegionalAnalysis}
+          aria-label={messages.analysis.openRegionalAnalysis}
+        >{analysis.name}</a>
+        : analysis.name}
       {!complete &&
         <span>
           <div className='progress'>
@@ -105,27 +89,6 @@ function Analysis ({analysis, onSelect, deleteRegionalAnalysis}) {
             {analysis.status.complete} of {analysis.status.total} complete
           </small>
         </span>}
-
-      <a
-        onClick={e => deleteRegionalAnalysis(analysis.id)}
-        tabIndex={0}
-        title={messages.analysis.deleteRegionalAnalysis}
-        aria-label={messages.analysis.deleteRegionalAnalysis}
-        className='pull-right'
-      >
-        <Icon type='trash' />
-      </a>
-
-      {complete &&
-        <a
-          onClick={e => onSelect(analysis.id)}
-          tabIndex={0}
-          title={messages.analysis.openRegionalAnalysis}
-          aria-label={messages.analysis.openRegionalAnalysis}
-          className='pull-right'
-        >
-          <Icon type='arrow-right' />
-        </a>}
     </li>
   )
 }
