+++ lib/containers/application.js
@@ -9,16 +9,13 @@ import selectActiveModification from '../selectors/active-modification'
 import selectBundleId from '../selectors/bundle-id'
 import selectProjectId from '../selectors/current-project-id'
 import selectScenarioId from '../selectors/current-scenario-id'
+import selectScenarioIsLoaded from '../selectors/scenario-is-loaded'
 import selectModificationBounds from '../selectors/modification-bounds'
 
 function mapStateToProps (state, props) {
-  const {mapState, network, scenario, user, analysis} = state
-  const {params} = props
+  const {mapState, network, user, analysis} = state
   const {error, outstandingRequests} = network
   const hasError = !!error
-  const {feeds} = scenario
-  const currentScenario = scenario.scenariosById[params.scenarioId]
-  const currentBundle = currentScenario && currentScenario.bundleId
 
   const scenarioId = selectScenarioId(state, props)
   return {
@@ -34,12 +31,7 @@ function mapStateToProps (state, props) {
     outstandingRequests,
     projectId: selectProjectId(state, props),
     scenarioId,
-    scenarioIsLoaded: !!(
-      currentBundle &&
-      currentScenario &&
-      feeds &&
-      feeds.length > 0
-    ),
+    scenarioIsLoaded: selectScenarioIsLoaded(state, props),
     username: user.profile && user.profile.name,
     zoom: mapState.zoom,
     analysisMode: analysis.active
