+++ lib/containers/bookmark-chooser.js
@@ -10,16 +10,17 @@ import {
 } from '../actions/analysis'
 import {createBookmark} from '../actions/project'
 import BookmarkChooser from '../components/analysis/bookmark-chooser'
+import selectCurrentBundle from '../selectors/current-bundle'
 import comparisonInProgress from '../selectors/comparison-in-progress'
 
-function mapStateToProps (state) {
+function mapStateToProps (state, ownProps) {
   const {project, analysis, scenario} = state
   return {
     bookmarks: project.currentProject.bookmarks,
     isochroneLonLat: analysis.isochroneLonLat,
     isFetchingIsochrone: analysis.isFetchingIsochrone,
     profileRequest: analysis.profileRequest,
-    bundleId: scenario.currentBundle.id,
+    bundleId: selectCurrentBundle(state, ownProps).id,
     isochroneCutoff: analysis.isochroneCutoff,
     modifications: scenario.modifications.filter(
       m => m.variants[analysis.activeVariant]
