+++ lib/containers/isochrone.js
@@ -14,11 +14,13 @@ import {
 import isochrone from '../selectors/isochrone'
 import comparisonIsochrone from '../selectors/comparison-isochrone'
 import comparisonInProgress from '../selectors/comparison-in-progress'
+import selectCurrentBundle from '../selectors/current-bundle'
 
 function mapStateToProps (state, props) {
   const {analysis, scenario, project, mapState} = state
+  const currentBundle = selectCurrentBundle(state, props)
   return {
-    bundleId: scenario.currentBundle.id,
+    bundleId: currentBundle.id,
     isFetchingIsochrone: analysis.isFetchingIsochrone,
     isochrone: isochrone(state),
     isochroneCutoff: analysis.isochroneCutoff,
