+++ lib/containers/map-control.js
@@ -5,10 +5,15 @@ import {addComponent, removeComponent, setCenter} from '../actions/map'
 import MapControl from '../components/map/control'
 import {ISOCHRONE_COMPONENT, OPPORTUNITY_COMPONENT} from '../constants/map'
 
+import selectCurrentBundle from '../selectors/current-bundle'
+import selectCurrentScenario from '../selectors/current-scenario'
+
 function mapStateToProps (state, props) {
   const {analysis, mapState, scenario} = state
+  const currentBundle = selectCurrentBundle(state, props)
+  const currentScenario = selectCurrentScenario(state, props)
   return {
-    bundleId: scenario.currentBundle ? scenario.currentBundle.id : null,
+    bundleId: currentBundle ? currentBundle.id : null,
     center: mapState.center,
     geocoderApiKey: process.env.MAPZEN_SEARCH_KEY,
     isochroneCutoff: analysis.isochroneCutoff,
@@ -17,7 +22,7 @@ function mapStateToProps (state, props) {
     isFetchingIsochrone: analysis.isFetchingIsochrone,
     isShowingIsochrone: mapState.components.indexOf(ISOCHRONE_COMPONENT) !== -1,
     modifications: scenario.modifications,
-    scenarioId: scenario.currentScenario ? scenario.currentScenario.id : null,
+    scenarioId: currentScenario ? currentScenario.id : null,
     accessibility: analysis.accessibility
   }
 }
