+++ lib/containers/regional-analysis-results.js
@@ -11,7 +11,7 @@ import {
   uploadAggregationArea
 } from '../actions/analysis/regional'
 import {addComponent, removeComponent} from '../actions/map'
-import RegionalAnalysis from '../components/analysis/regional'
+import RegionalAnalysis from '../components/analysis/regional-results'
 import {
   REGIONAL_COMPONENT,
   REGIONAL_COMPARISON_COMPONENT,
@@ -24,16 +24,10 @@ import selectComparisonAggregateAccessibility from '../selectors/comparison-aggr
 
 function mapStateToProps (state, ownProps) {
   const {project, analysis} = state
-  const {regionalAnalysisId} = ownProps.params
 
   return {
     regionalAnalyses: project.regionalAnalyses,
     projectId: project.currentProject.id,
-    analysisId: regionalAnalysisId, // needed so we can request grids before analyses have loaded when deep linking
-    analysis:
-      project.regionalAnalyses != null
-        ? project.regionalAnalyses.find(a => a.id === regionalAnalysisId)
-        : null,
     comparisonAnalysis:
       project.regionalAnalyses != null
         ? project.regionalAnalyses.find(
