+++ lib/containers/single-point-analysis.js
@@ -19,9 +19,7 @@ import {
   setProfileRequest
 } from '../actions/analysis'
 import {
-  runAnalysis,
-  load as loadRegionalAnalyses,
-  deleteRegionalAnalysis,
+  createRegionalAnalysis,
   setRegionalAnalysisBounds
 } from '../actions/analysis/regional'
 import {addComponent, removeComponent} from '../actions/map'
@@ -30,6 +28,8 @@ import comparisonAccessibility from '../selectors/comparison-accessibility'
 import percentileCurves from '../selectors/percentile-curves'
 import comparisonPercentileCurves from '../selectors/comparison-percentile-curves'
 import comparisonInProgress from '../selectors/comparison-in-progress'
+import selectCurrentBundle from '../selectors/current-bundle'
+import selectRegionalAnalyses from '../selectors/regional-analyses'
 
 function mapStateToProps (state, ownProps) {
   const {analysis, mapState, scenario, project} = state
@@ -40,10 +40,12 @@ function mapStateToProps (state, ownProps) {
   const scenarioId = scenario.currentScenario
     ? scenario.currentScenario.id
     : null
+
+  const currentBundle = selectCurrentBundle(state, ownProps)
   return {
     // TODO duplicate code below and in containers/isochrone
     accessibility: accessibility(state, ownProps),
-    bundleId: scenario.currentBundle ? scenario.currentBundle.id : null,
+    bundleId: currentBundle ? currentBundle.id : null,
     comparisonAccessibility: comparisonAccessibility(state, ownProps),
     comparisonBundleId: analysis.comparisonBundleId,
     comparisonInProgress: comparisonInProgress(state),
@@ -64,15 +66,7 @@ function mapStateToProps (state, ownProps) {
     modifications: scenario.modifications.filter(m => m.variants[variantId]),
     profileRequest: analysis.profileRequest,
     projectId: project.currentProject.id,
-    regionalAnalyses:
-      project.regionalAnalyses != null
-        ? project.regionalAnalyses.filter(
-            a =>
-              a.scenarioId === scenarioId &&
-              a.variant === variantId &&
-              !a.deleted
-          )
-        : undefined,
+    regionalAnalyses: selectRegionalAnalyses(state, ownProps),
     regionalAnalysisBounds: analysis.regional.bounds,
     projectBounds: project.currentProject.bounds,
     scenarioApplicationErrors: analysis.scenarioApplicationErrors,
@@ -92,8 +86,6 @@ function mapDispatchToProps (dispatch, props) {
   return {
     clearComparisonScenario: () => dispatch(clearComparisonScenario()),
     clearIsochroneResults: () => dispatch(clearIsochroneResults()),
-    deleteRegionalAnalysis: analysisId =>
-      dispatch(deleteRegionalAnalysis(analysisId)),
     enterAnalysisMode: () => dispatch(enterAnalysisMode()),
     exitAnalysisMode: () => dispatch(exitAnalysisMode()),
     fetchTravelTimeSurface: opts =>
@@ -103,8 +95,6 @@ function mapDispatchToProps (dispatch, props) {
           dispatch
         })
       ),
-    loadRegionalAnalyses: projectId =>
-      dispatch(loadRegionalAnalyses(projectId)),
     push: url => dispatch(push(url)),
     addEditRegionalAnalysisBoundsLayerToMap: () =>
       dispatch(addComponent(EDIT_REGIONAL_ANALYSIS_BOUNDS_COMPONENT)),
@@ -117,11 +107,7 @@ function mapDispatchToProps (dispatch, props) {
       dispatch(removeComponent(ISOCHRONE_COMPONENT)),
     removeOpportunityLayerFromMap: () =>
       dispatch(removeComponent(OPPORTUNITY_COMPONENT)),
-    runAnalysis: request => dispatch(runAnalysis(request)),
-    selectRegionalAnalysis: (projectId, scenarioId, regionalAnalysisId) =>
-      dispatch(
-        push(`/scenarios/${scenarioId}/analysis/regional/${regionalAnalysisId}`)
-      ),
+    createRegionalAnalysis: request => dispatch(createRegionalAnalysis(request)),
     setActiveVariant: v => dispatch(setActiveVariant(v)),
     setComparisonScenario: ({id, bundleId}, variantIndex) =>
       dispatch(setComparisonScenario({id, bundleId, variantIndex})),
