+++ lib/routes.js
@@ -23,6 +23,7 @@ import SelectScenario from './containers/select-scenario'
 import SinglePointAnalysis from './containers/single-point-analysis'
 import CreateGrid from './containers/create-grid'
 import Report from './containers/report'
+import RegionalPage from './containers/regional'
 import Regional from './containers/regional-analysis-results'
 import {setStore} from './utils/authenticated-fetch'
 
@@ -67,6 +68,8 @@ export default function Routes ({
           />
           <Route path='analysis' component={SinglePointAnalysis} />
           <Route path='analysis/:variantId' component={SinglePointAnalysis} />
+          <Route path='regional' component={RegionalPage} />
+          <Route path='regional/:regionalAnalysisId' component={RegionalPage} />
           <Route
             path='analysis/regional/:regionalAnalysisId'
             component={Regional}
