+++ lib/selectors/active-regional-analysis.js
@@ -1,0 +1,8 @@
+import {createSelector} from 'reselect'
+
+import selectRegionalAnalyses from './regional-analyses'
+
+export default createSelector(
+  [selectRegionalAnalyses, (state, props) => props.params.regionalAnalysisId],
+  (analyses, id) => analyses.find((a) => a.id === id)
+)
