+++ lib/styles.css
@@ -170,6 +170,7 @@ input[type="number"] {
   left: 40px;
   height: 100%;
   border-right: 1px solid #ddd;
+  overflow: hidden;
 }
 
 .ApplicationDock.analysisMode,
